% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{conv_round}
\alias{conv_round}
\title{Function to round cleanly}
\usage{
conv_round(x)
}
\arguments{
\item{x}{single number to round}
}
\value{
rounded x
}
\description{
Function to round cleanly
}
