% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtDataLoadFuncs.R
\name{load_county_pops}
\alias{load_county_pops}
\title{Function to load dataframe of county-geoid crosswalk}
\usage{
load_county_pops(region_code, state_fips, pop_year)
}
\arguments{
\item{region_code}{}

\item{shapefile_fn}{}
}
\value{
Grab county names and population with geoid
}
\description{
Function to load dataframe of county-geoid crosswalk
}
\author{
Elizabeth Lee
}
