% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData.R
\name{load_cum_inf_geounit_dates}
\alias{load_cum_inf_geounit_dates}
\title{Convenience function to load cumulative geounit infections at multiple specific dates for the given scenarios}
\usage{
load_cum_inf_geounit_dates(
  scn_dirs,
  display_dates = config$report$formatting$display_dates,
  num_files = NA,
  scenariolabels = NULL,
  incl_geoids = NULL,
  geoid_len = 0,
  padding_char = "0",
  file_extension = "auto"
)
}
\arguments{
\item{scn_dirs}{paste(config$name, config$interventions$scenarios, sep = "_") character vector of scenario directory names}

\item{display_dates}{config$report$formatting$display_dates character vector of dates to for which cumulative infections should be extracted}

\item{scenariolabels}{config$report$formatting$scenario_labels character vector of scenario labels}

\item{incl_geoids}{character vector of geoids that are included in the report}

\item{geoid_len}{in defined, this we want to make geoids all the same length}

\item{padding_char}{character to add to the front of geoids if fixed length}
}
\value{
a data frame with columns
         - time
         - comp
         - geoid
         - N
         - sim_num
         - scenario_num
         - scenario_name
}
\description{
Convenience function to load cumulative geounit infections at multiple specific dates for the given scenarios
}
