% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData.R
\name{load_geodata_file}
\alias{load_geodata_file}
\title{Convenience function to load geodata.csv}
\usage{
load_geodata_file(filename, geoid_len = 0, geoid_pad = "0", to_lower = FALSE)
}
\arguments{
\item{filename}{geodata.csv filename}

\item{geoid_len}{length of geoid character string}

\item{geoid_pad}{what to pad the geoid character string with}

\item{to_lower}{whether to make all column names lowercase}
}
\value{
a data frame with columns
        -
}
\description{
Convenience function to load geodata.csv
}
