% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData.R
\name{load_hosp_geocombined_totals}
\alias{load_hosp_geocombined_totals}
\title{Convenience function to allow us to load hospital totals for the combined geounits quickly for
the given scenarios.}
\usage{
load_hosp_geocombined_totals(
  scn_dirs,
  num_files = NA,
  scenariolabels = NULL,
  name_filter,
  incl_geoids = NULL,
  geoid_len = 0,
  padding_char = "0",
  file_extension = "auto"
)
}
\arguments{
\item{scn_dirs}{the dirctories containing the relevant scenarios}

\item{scenariolabels}{the scenarios labels}

\item{name_filter}{character string that filenames should match}

\item{geoid_len}{required length of geoid}

\item{padding_char}{padding}
}
\value{
a data frame with columns:
   - sim_num
   - NhospCurr number of people in hospital on a day
   - NICUCurr number of people in ICU on a day
   - NincidDeath  number of incidence deaths on a day
   - NincidInf  number of incident infections on a day
   - NincidICH number of incident ICUs on a day
}
\description{
Convenience function to allow us to load hospital totals for the combined geounits quickly for
the given scenarios.
}
