% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData.R
\name{load_hosp_geounit_relative_to_threshold}
\alias{load_hosp_geounit_relative_to_threshold}
\title{Convenience function to load hosp data and display log proportion}
\usage{
load_hosp_geounit_relative_to_threshold(
  scn_dirs,
  threshold,
  variable,
  end_date = config$end_date,
  name_filter,
  num_files = NA,
  incl_geoids = NULL,
  scenario_labels = NULL,
  geoid_len = 0,
  padding_char = "0",
  file_extension = "auto"
)
}
\arguments{
\item{scn_dirs}{paste(config$name, config$interventions$scenarios, sep = "_") character vector of scenario directory names}

\item{threshold}{A named numeric vector where the names are geoids and values are thresholds}

\item{variable}{character string of variable to which to compare threshold}

\item{end_date}{simulation end date character string}

\item{name_filter}{pdeath character string (high, med, low)}

\item{num_files}{number of files to read in or NA}

\item{incl_geoids}{optional character vector of geoids that are included in the report, if not included, all geoids will be used}

\item{scenario_labels}{config$report$formatting$scenario_labels character vector of scenario labels}

\item{geoid_len}{required length of geoid}

\item{padding_char}{padding}
}
\value{
a data frame with columns
        - scenario_name
        - geoid
        - time
        - NhospCurr number of people in hospital on a day
        - NICUCurr number of people in ICU on a day
        - NVentCurr  number of ventilators used on a day
        - threshold_value numeric values from threshold named vector
        - prop_needed ratio of needed beds/ventilators relative to threshold value
        - log_prop_needed log ratio of needed beds/ventilators with plotting edits on the borders
}
\description{
Convenience function to load hosp data and display log proportion
}
