% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData.R
\name{load_hosp_geounit_threshold}
\alias{load_hosp_geounit_threshold}
\title{Convenience function to load the slice for each geoid where the value of an outcome exceeds a given threshold}
\usage{
load_hosp_geounit_threshold(
  scn_dirs,
  threshold,
  variable,
  end_date = config$end_date,
  name_filter,
  num_files = NA,
  incl_geoids = NULL,
  scenario_labels = NULL,
  geoid_len = 0,
  padding_char = "0",
  file_extension = "auto"
)
}
\arguments{
\item{scn_dirs}{paste(config$name, config$interventions$scenarios, sep = "_") character vector of scenario directory names}

\item{threshold}{A named numeric vector.  This function will pull the first time slice that meets or exceeds all thresholds}

\item{variable}{character string of variable to which to compare threshold}

\item{end_date}{simulation end date character string}

\item{incl_geoids}{optional character vector of geoids that are included in the report, if not included, all geoids will be used}

\item{scenario_labels}{config$report$formatting$scenario_labels character vector of scenario labels}

\item{geoid_len}{required length of geoid}

\item{padding_char}{padding}
}
\value{
a data frame with columns
        - sim_num
        - NhospCurr number of people in hospital on a day
        - NICUCurr number of people in ICU on a day
        - NincidDeath  number of incidence deaths on a day
        - NincidInf  number of incident infections on a day
        - NincidICH number of incident ICUs on a day
}
\description{
Convenience function to load the slice for each geoid where the value of an outcome exceeds a given threshold
}
