% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadFuncs.R
\name{load_hosp_sims_filtered}
\alias{load_hosp_sims_filtered}
\title{Function loads multiple hospital simulations into a combined data frame
with pre and post filters}
\usage{
load_hosp_sims_filtered(
  scenario_dir,
  pdeath = "low",
  post_process = function(x) {     x }
)
}
\arguments{
\item{scenario_dir}{the subdirectory containing this scenario}

\item{pdeath}{probablity of death for these simulations....probably can be reduced.}

\item{post_process}{function that does processing after}
}
\description{
Function loads multiple hospital simulations into a combined data frame
with pre and post filters
}
\author{
Justin Lessler
}
