% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadFuncs.R
\name{load_scenario_sims_filtered}
\alias{load_scenario_sims_filtered}
\title{Function designed to allow generic filtering and summarization
of simulations before merging them together when loading outputs
from the infection genertion modeling pipeline.}
\usage{
load_scenario_sims_filtered(
  scenario_dir,
  num_files = NA,
  post_process = function(x) {     x },
  pre_process = function(x) {     x },
  geoid_len = 0,
  padding_char = "0",
  file_extension = "auto",
  ...
)
}
\arguments{
\item{scenario_dir}{the subdirectory containing this scenario}

\item{post_process}{function that does processing after}

\item{geoid_len}{in defined, this we want to make geoids all the same length}

\item{padding_char}{character to add to the front of geoids if fixed length}

\item{...}{additional parameters to pass to pre and/or post process}
}
\value{
a combined data frame of all hospital simulations with filters applied pre merge.
}
\description{
Function designed to allow generic filtering and summarization
of simulations before merging them together when loading outputs
from the infection genertion modeling pipeline.
}
\author{
Justin Lessler
}
