% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtDataLoadFuncs.R
\name{load_shapefile}
\alias{load_shapefile}
\title{Function to load shapefile for report location.}
\usage{
load_shapefile(region_code, state_fips, shapefile_fn)
}
\arguments{
\item{scenario_dir}{the subdirectory containing this scenario}

\item{post_process}{function that does processing after}
}
\value{
shapefile for region of interest with appropriate administrative units
}
\description{
Function to load shapefile for report location.
}
\author{
Elizabeth Lee
}
