% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData.R
\name{load_ts_current_hosp_geounit}
\alias{load_ts_current_hosp_geounit}
\title{Convenience function to load timeseries current hospital outcomes}
\usage{
load_ts_current_hosp_geounit(
  scn_dirs,
  num_files = NA,
  scenariolabels = NULL,
  name_filter,
  end_date,
  incl_geoids = NULL,
  geoid_len = 0,
  padding_char = "0",
  qlo = 0.025,
  qhi = 0.975,
  file_extension = "auto"
)
}
\arguments{
\item{scn_dirs}{paste(config$name, config$interventions$scenarios, sep = "_") character vector of scenario directory names}

\item{scenariolabels}{config$report$formatting$scenario_labels character vector of scenario labels}

\item{name_filter}{character string that filenames should match}

\item{end_date}{last date to include in timeseries}

\item{incl_geoids}{character vector of geoids that are included in the report}

\item{geoid_len}{in defined, this we want to make geoids all the same length}

\item{padding_char}{character to add to the front of geoids if fixed length}
}
\value{
a data frame with columns
         - time
         - geoid
         - NHospCurr, NICUCurr, NVentCurr
         - sim_num
         - scenario_num
         - scenario_name
}
\description{
Convenience function to load timeseries current hospital outcomes
}
