% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{make_excess_heatmap}
\alias{make_excess_heatmap}
\title{Plot heatmap of excess variables, as absolute number
(default), per-capita number needed, or % over threshold}
\usage{
make_excess_heatmap(
  hosp_dat,
  current_scenario,
  type,
  threshold,
  shp,
  varname,
  varlabel,
  geodata = NULL,
  popnodes = config$spatial_setup$popnodes,
  trunc_value = NULL
)
}
\arguments{
\item{hosp_dat}{timeseries with incident and cumulative hospitalizations, ICU admissions, etc.}

\item{current_scenario}{text string of scenario label to plot}

\item{type}{either "absolute", for total number excess; "percent", for percent excess; or "per-capita" for number excess per 10,000 population}

\item{threshold}{named vector of threshold value for given variable}

\item{shp}{shapefile with GEOID names}

\item{varname}{character string of variable name to plot from hosp data}

\item{varlabel}{character string of variable name for plot legend}

\item{geodata}{geodata object from load_config; only needed if type="per-capita"}

\item{popnodes}{name of pop variable in geodata; only needed if type="per-capita"}

\item{trunc_value}{value at which to truncate values for easy plotting; if null, defaults to max}
}
\description{
Plot heatmap of excess variables, as absolute number
(default), per-capita number needed, or % over threshold
}
