% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{make_scn_state_table_cap}
\alias{make_scn_state_table_cap}
\title{Make caption for statewide table of infections, hosp, ICU, deaths for given scenario}
\usage{
make_scn_state_table_cap(current_scenario, table_date_labels, table_num = "")
}
\arguments{
\item{current_scenario}{text string of scenario label for which to build table}
}
\value{
plot of distribution of peak timing across simulations by county
}
\description{
Make caption for statewide table of infections, hosp, ICU, deaths for given scenario
}
