% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_hist_incidHosp_state}
\alias{plot_hist_incidHosp_state}
\title{Plot figure showing histogram of cumulative hospitalizations by a certain date}
\usage{
plot_hist_incidHosp_state(
  hosp_state_totals,
  pdeath_level = "high",
  sim_start_date,
  summary_date
)
}
\arguments{
\item{hosp_state_totals}{totals for hospitalization related data for state for all pdeath}

\item{pdeath_level}{level of IFR (string: high/med/low) for filtering hospitalization data}

\item{sim_start_date}{simulation start date as character string "2020-01-01"}

\item{summary_date}{date at which to present cumulative summary of hospitalizations}
}
\value{
plot of cum hosp for state across simulations
}
\description{
Plot figure showing histogram of cumulative hospitalizations by a certain date
}
