% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_line_hospPeak_time_county}
\alias{plot_line_hospPeak_time_county}
\title{Plot figure showing histogram of peak hospital occupancy by a certain date}
\usage{
plot_line_hospPeak_time_county(
  hosp_cty_peaks,
  cty_names,
  pdeath_level = "high",
  scenario_labels,
  scenario_cols,
  start_date,
  end_date
)
}
\arguments{
\item{hosp_cty_peaks}{totals for hospitalization related data for state for all pdeath}

\item{cty_names}{Dataframe with county geoid and county name}

\item{pdeath_level}{level of IFR (string: high/med/low) for filtering hospitalization data}

\item{scenario_labels}{scenario names from config}

\item{scenario_cols}{scenario colors from config}

\item{start_date}{date to filter to start search for peak timing (character string)}

\item{end_date}{date to filter to end search for peak timing (character string)}
}
\value{
plot of distribution of peak timing across simulations by county
}
\description{
Plot figure showing histogram of peak hospital occupancy by a certain date
}
