% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_model_vs_obs}
\alias{plot_model_vs_obs}
\title{Compare model outputs and data from CSSE}
\usage{
plot_model_vs_obs(
  state_hosp_totals,
  jhu_obs_dat,
  scenario_labels,
  scenario_cols,
  pdeath_level,
  obs_data_col = "black",
  ci.L = 0,
  ci.U = 1,
  date_breaks = "1 month",
  sim_start_date,
  sim_end_date,
  assumed_reporting_rate
)
}
\arguments{
\item{state_hosp_totals}{state hosp data frame}

\item{jhu_obs_dat}{dataframe with CSSE data}

\item{scenario_labels}{character vector with scenario labels}

\item{scenario_cols}{character vector with scenario colors}

\item{pdeath_level}{IFR level assumption}

\item{obs_data_col}{character string of observed data color}

\item{ci.L}{lower bound confidence interval}

\item{ci.U}{upper bound confidence interval}

\item{date_breaks}{breaks for dates in figure}

\item{sim_start_date}{simulation start date}

\item{sim_end_date}{simulation end date}

\item{assumed_reporting_rate}{assumed reporting rate (0.2 means 20% of infections are reported cases)}
}
\description{
Compare model outputs and data from CSSE
}
