% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_ts_incid_death_state_sample_allPdeath}
\alias{plot_ts_incid_death_state_sample_allPdeath}
\title{Plot figure showing 15 random sims of incident infections}
\usage{
plot_ts_incid_death_state_sample_allPdeath(
  hosp_state_totals,
  num_sims = 15,
  sim_start_date,
  sim_end_date,
  plot_intervention = FALSE,
  interv_start_date = NA,
  interv_end_date = NA
)
}
\arguments{
\item{hosp_state_totals}{totals for hospitalization related data for state for all pdeath}

\item{num_sims}{the number of simulations to show}

\item{sim_start_date}{simulation start date as character string "2020-01-01"}

\item{sim_end_date}{simulation end date as character string}

\item{plot_intervention}{logical indicating whether to plot grey box over a single intervention period -- will need to adapt if we want to show multiple intervention periods}

\item{interv_start_date}{intervention start date as character string}

\item{interv_end_date}{intervention end date as character string}

\item{pdeath_level}{level of IFR (string: high/med/low) for filtering hospitalization data}
}
\value{
plot with 15 random simulations of incident deaths
}
\description{
Plot figure showing 15 random sims of incident infections
}
