% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{print_pretty_date_short}
\alias{print_pretty_date_short}
\title{Pretty print short date for text and tables}
\usage{
print_pretty_date_short(date_string)
}
\value{
a function that formats short pretty dates
}
\description{
Pretty print short date for text and tables
}
