% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadFuncs.R
\name{load_hosp_sims_filtered}
\alias{load_hosp_sims_filtered}
\title{Function loads multiple hospital simulations into a combined data frame
with pre and post filters}
\usage{
load_hosp_sims_filtered(
  scenario_dir,
  name_filter = "",
  loc_filter = state_fips,
  random_sample_filter = NA,
  post_process = function(x) {     x }
)
}
\arguments{
\item{scenario_dir}{the subdirectory containing this scenario}

\item{name_filter}{function that}

\item{post_process}{function that does processing after}
}
\value{
a combined data frame of all hospital simulations with filters applied pre merge.
}
\description{
Function loads multiple hospital simulations into a combined data frame
with pre and post filters
}
\author{
Justin Lessler
}
