% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadFuncs.R
\name{load_hosp_sims_filtered}
\alias{load_hosp_sims_filtered}
\title{Function loads multiple hospital simulations into a combined data frame
with pre and post filters}
\usage{
load_hosp_sims_filtered(
  scenario_dir,
  name_filter,
  num_files = NA,
  post_process = function(x) {     x },
  geoid_len = 0,
  padding_char = "0",
  file_extension = "auto",
  ...
)
}
\arguments{
\item{scenario_dir}{the subdirectory containing this scenario}

\item{name_filter}{function that}

\item{post_process}{function that does processing after}

\item{geoid_len}{in defined, this we want to make geoids all the same length}

\item{padding_char}{character to add to the front of geoids if fixed length}

\item{...}{additional parameters to post process function}
}
\value{
a combined data frame of all hospital simulations with filters applied pre merge.
}
\description{
Function loads multiple hospital simulations into a combined data frame
with pre and post filters
}
\author{
Justin Lessler
}
