% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadFuncs.R
\name{load_scenario_sims_filtered}
\alias{load_scenario_sims_filtered}
\title{Function designed to allow generic filtering and summarization
of simulations before merging them together when loading outputs
from the infection genertion modeling pipeline.}
\usage{
load_scenario_sims_filtered(
  scenario_dir,
  loc_filter = state_fips,
  post_process = function(x) {     x },
  pre_process = function(x) {     x }
)
}
\arguments{
\item{scenario_dir}{the subdirectory containing this scenario}

\item{post_process}{function that does processing after}
}
\value{
a combined data frame of all hospital simulations with filters applied pre merge.
}
\description{
Function designed to allow generic filtering and summarization
of simulations before merging them together when loading outputs
from the infection genertion modeling pipeline.
}
\author{
Justin Lessler
}
