% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{make_scn_state_table}
\alias{make_scn_state_table}
\title{Make statewide table of infections, hosp, ICU, deaths for given scenario}
\usage{
make_scn_state_table(current_scenario, hosp_state_totals, table_dates, params)
}
\arguments{
\item{current_scenario}{text string of scenario label for which to build table}

\item{hosp_state_totals}{totals for hospitalization related data for state for all pdeath}

\item{table_dates}{formatted table_dates object}

\item{params}{parameter object with table_dates, table_date_labels, pdeath_labels, pdeath_filecode}
}
\value{
plot of distribution of peak timing across simulations by county
}
\description{
Make statewide table of infections, hosp, ICU, deaths for given scenario
}
