% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_line_hospPeak_time_county}
\alias{plot_line_hospPeak_time_county}
\title{THIS FUNCTION HAS NOT YET BEEN TESTED 3/25/2020 ECL}
\usage{
plot_line_hospPeak_time_county(
  hosp_cty_peaks,
  cty_names,
  pdeath_level = "high",
  start_date,
  end_date
)
}
\arguments{
\item{pdeath_level}{level of IFR (string: high/med/low) for filtering hospitalization data}

\item{start_date}{date to filter to start search for peak timing (character string)}

\item{end_date}{date to filter to end search for peak timing (character string)}

\item{hosp_state_totals}{totals for hospitalization related data for state for all pdeath}
}
\value{
plot of distribution of peak timing across simulations by county
}
\description{
Plot figure showing histogram of peak hospital occupancy by a certain date
}
