% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_ts_hosp_state_sample}
\alias{plot_ts_hosp_state_sample}
\title{Plot figure showing 15 random sims of hospitalization occupancy}
\usage{
plot_ts_hosp_state_sample(
  hosp_state_totals,
  varname = "NhospCurr",
  varlabel = "Daily hospital occupancy",
  num_sims = 15,
  pdeath_level = "high",
  scenario_labels,
  scenario_cols,
  sim_start_date,
  sim_end_date,
  plot_intervention = FALSE,
  interv_start_date = NA,
  interv_end_date = NA
)
}
\arguments{
\item{hosp_state_totals}{totals for hospitalization related data for state for all pdeath}

\item{varname}{character string of variable name to plot from hosp data}

\item{varlabel}{character string of varialbe name label for plot}

\item{num_sims}{the number of simulations to show}

\item{pdeath_level}{level of IFR for filtering hospitalization data - TODO: Move our of functions}

\item{scenario_labels}{names of the scenarios to include- TODO: give a default}

\item{scenario_cols}{colors to plot each scenario in  - TODO: give a default}

\item{sim_start_date}{simulation start date as character string "2020-01-01"}

\item{sim_end_date}{simulation end date as character string}

\item{plot_intervention}{logical indicating whether to plot grey box over a single intervention period -- will need to adapt if we want to show multiple intervention periods}

\item{interv_start_date}{intervention start date as character string}

\item{interv_end_date}{intervention end date as character string}
}
\value{
plot with N random simulations of hospital occupancy
}
\description{
TODO ADD OPTION TO CHANGE VARIABLE THAT IS PLOTTED TO ANYTHING IN HOSP OUTCOMES DATASET
}
